move cache logic to internal
